+++ app/src/main/java/ua/com/radiokot/camerapp/about/ui/AboutScreen.kt
@@ -28,7 +28,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeContentPadding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.BasicText
@@ -68,6 +67,7 @@ import ua.com.radiokot.camerapp.ui.LocalColors
 import ua.com.radiokot.camerapp.ui.PodkovaFamily
 import ua.com.radiokot.camerapp.ui.Vignette
 import ua.com.radiokot.camerapp.ui.paperBackground
+import ua.com.radiokot.camerapp.util.barsAndCutoutPadding
 
 @Composable
 fun AboutScreen(
@@ -80,7 +80,7 @@ fun AboutScreen(
             state = rememberScrollState(),
             overscrollEffect = null,
         )
-        .safeContentPadding()
+        .barsAndCutoutPadding()
         .padding(24.dp)
 ) {
     val colors = LocalColors.current