+++ app/src/main/java/ua/com/radiokot/camerapp/discardchanges/ui/ConfirmDiscardChangesDialog.kt
@@ -27,7 +27,6 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeContentPadding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.BasicText
 import androidx.compose.runtime.Composable
@@ -42,6 +41,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import ua.com.radiokot.camerapp.ui.LocalColors
 import ua.com.radiokot.camerapp.ui.PodkovaFamily
+import ua.com.radiokot.camerapp.util.barsAndCutoutPadding
 
 @Composable
 fun ConfirmDiscardChangesDialog(
@@ -50,7 +50,7 @@ fun ConfirmDiscardChangesDialog(
 ) = Box(
     contentAlignment = Alignment.Center,
     modifier = Modifier
-        .safeContentPadding()
+        .barsAndCutoutPadding()
 ) {
     val colors = LocalColors.current
     val actionTextStyle = remember(colors) {