+++ app/src/main/java/ua/com/radiokot/camerapp/envelopes/ui/SaveEnvelopeStampsScreen.kt
@@ -26,7 +26,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeContentPadding
 import androidx.compose.foundation.text.BasicText
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.FloatState
@@ -43,6 +42,7 @@ import ua.com.radiokot.camerapp.ui.LocalColors
 import ua.com.radiokot.camerapp.ui.PodkovaFamily
 import ua.com.radiokot.camerapp.ui.ProgressBar
 import ua.com.radiokot.camerapp.ui.paperBackground
+import ua.com.radiokot.camerapp.util.barsAndCutoutPadding
 
 /**
  * @param progressState value in [0..1] range
@@ -55,7 +55,7 @@ fun SaveEnvelopeStampsScreen(
     verticalArrangement = Arrangement.Center,
     horizontalAlignment = Alignment.CenterHorizontally,
     modifier = modifier
-        .safeContentPadding()
+        .barsAndCutoutPadding()
         .padding(24.dp),
 ) {
     BasicText(