+++ app/src/main/java/ua/com/radiokot/camerapp/intro/ui/PermissionsScreen.kt
@@ -29,7 +29,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeContentPadding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.BasicText
@@ -55,6 +54,7 @@ import ua.com.radiokot.camerapp.ui.LocalColors
 import ua.com.radiokot.camerapp.ui.PodkovaFamily
 import ua.com.radiokot.camerapp.ui.Vignette
 import ua.com.radiokot.camerapp.ui.paperBackground
+import ua.com.radiokot.camerapp.util.barsAndCutoutPadding
 
 @Composable
 fun PermissionsScreen(
@@ -64,7 +64,7 @@ fun PermissionsScreen(
     onGrantAction: () -> Unit,
 ) = Column(
     modifier = modifier
-        .safeContentPadding()
+        .barsAndCutoutPadding()
         .padding(24.dp),
 ) {
     val permissionSet = remember(permissions) {