+++ app/src/main/java/ua/com/radiokot/camerapp/posters/ui/EditStampPosterTextDialog.kt
@@ -34,7 +34,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.safeContentPadding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.BasicText
@@ -68,6 +67,7 @@ import ua.com.radiokot.camerapp.ui.AppTheme
 import ua.com.radiokot.camerapp.ui.LeField
 import ua.com.radiokot.camerapp.ui.LocalColors
 import ua.com.radiokot.camerapp.ui.rememberLeFieldTextStyle
+import ua.com.radiokot.camerapp.util.barsAndCutoutPadding
 
 @OptIn(ExperimentalLayoutApi::class)
 @Composable
@@ -84,7 +84,7 @@ fun EditStampPosterTextDialog(
             verticalAlignment = Alignment.CenterVertically,
             modifier = Modifier
                 .fillMaxSize()
-                .safeContentPadding()
+                .barsAndCutoutPadding()
                 .imePadding()
                 .clipToBounds()
         ) {
@@ -103,7 +103,7 @@ fun EditStampPosterTextDialog(
             horizontalAlignment = Alignment.CenterHorizontally,
             modifier = Modifier
                 .fillMaxSize()
-                .safeContentPadding()
+                .barsAndCutoutPadding()
                 .imePadding()
                 .clipToBounds()
         ) {