+++ app/src/main/java/ua/com/radiokot/camerapp/util/Insets.kt
@@ -1,0 +1,41 @@
+/* Copyright 2026 Oleg Koretsky
+
+   This file is part of the Press-Cut,
+   a digital postage stamp cutter Android app.
+
+   Press-Cut is free software: you can redistribute it
+   and/or modify it under the terms of the GNU General Public License
+   as published by the Free Software Foundation, either version 3 of the License,
+   or (at your option) any later version.
+
+   Press-Cut is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+   See the GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with Press-Cut. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+package ua.com.radiokot.camerapp.util
+
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.add
+import androidx.compose.foundation.layout.displayCutout
+import androidx.compose.foundation.layout.displayCutoutPadding
+import androidx.compose.foundation.layout.systemBars
+import androidx.compose.foundation.layout.systemBarsPadding
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+/**
+ * Use this instead of safe content,
+ * as safe content adds enormous gesture padding
+ * when the gesture navigation is used.
+ */
+fun Modifier.barsAndCutoutPadding() =
+    systemBarsPadding().displayCutoutPadding()
+
+val WindowInsets.Companion.barsAndCutout: WindowInsets
+    @Composable
+    get() = WindowInsets.systemBars.add(WindowInsets.displayCutout)